refactor: move util functions only used by one file into that file
