+++ src/create-client.js
@@ -1,7 +1,6 @@
 /* globals fetch */
 import 'isomorphic-fetch'
 import { API_BASE_URL } from './constants'
-import { buildUrlWithParams } from './utils'
 
 export const clientConstructor = ({ token, ...options }) => {
   return {
@@ -37,4 +36,13 @@ export const clientConstructor = ({ token, ...options }) => {
         .then(response => response.json())
     }
   }
+}
+
+export const buildUrlWithParams = (url, params = {}) => {
+  const queryParams = Object.keys(params)
+    .filter((key) => params[key] !== undefined && params[key] !== null)
+    .map((key) => `${encodeURIComponent(key)}=${encodeURIComponent(params[key])}`)
+    .join('&')
+
+  return queryParams ? `${url}?${queryParams}` : url
 }
