+++ src/utils.js
@@ -1,12 +1,3 @@
-export const buildUrlWithParams = (url, params = {}) => {
-  const queryParams = Object.keys(params)
-    .filter((key) => params[key] !== undefined && params[key] !== null)
-    .map((key) => `${encodeURIComponent(key)}=${encodeURIComponent(params[key])}`)
-    .join('&')
-
-  return queryParams ? `${url}?${queryParams}` : url
-}
-
 export const createMemberPatchQuery = ({ members, operation }) => {
   return members.map(member => ({
     op: operation,
@@ -15,25 +6,6 @@ export const createMemberPatchQuery = ({ members, operation }) => {
       email: member
     }
   }))
-}
-
-export const createOrUpdateWebhook = (http, { uid, tag, url, enable = false } = {}) => {
-  if (!url) {
-    throw new Error(`Please provide an url for ${tag}`)
-  }
-
-  if (!tag) {
-    throw new Error(`Please provide a tag name for the webhook`)
-  }
-
-  return http.request({
-    method: 'put',
-    url: `/forms/${uid}/webhooks/${tag}`,
-    data: {
-      url,
-      enable
-    }
-  })
 }
 
 export const isMemberPropValid = members => {
