+++ src/webhooks.js
@@ -1,5 +1,3 @@
-import { createOrUpdateWebhook } from './utils'
-
 export default http => new Webhooks(http)
 
 class Webhooks {
@@ -28,4 +26,23 @@ class Webhooks {
   update (args) {
     return createOrUpdateWebhook(this._http, args)
   }
+}
+
+const createOrUpdateWebhook = (http, { uid, tag, url, enable = false } = {}) => {
+  if (!url) {
+    throw new Error(`Please provide an url for ${tag}`)
+  }
+
+  if (!tag) {
+    throw new Error(`Please provide a tag name for the webhook`)
+  }
+
+  return http.request({
+    method: 'put',
+    url: `/forms/${uid}/webhooks/${tag}`,
+    data: {
+      url,
+      enable
+    }
+  })
 }
